% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{calculate_auc}
\alias{calculate_auc}
\title{calculate_auc}
\usage{
calculate_auc(test.y, pred.prob)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}
}
\value{
auc
}
\description{
Compute auc from predictions and truth
}

