% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration_plot}
\alias{calibration_plot}
\title{calibration_plot}
\usage{
calibration_plot(test.y, pred.prob)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}
}
\description{
Returns a ggplot2 plot object containing a smoothed propensity @ prediction level plot
}

