% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\docType{data}
\name{classifier_theme}
\alias{classifier_theme}
\title{classifier_theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 0.}
\usage{
classifier_theme
}
\description{
A useful theme for classifier diagnostic plots
}
\keyword{datasets}

