% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_plots.R
\name{classifierplots}
\alias{classifierplots}
\title{classifier.plots}
\usage{
classifierplots(test.y, pred.prob)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}
}
\description{
Produce a suit of classifier diagnostic plots
}
\examples{
\dontrun{
 classifierplots(example_predictions$test.y, example_predictions$pred.prob) 
}
}

