% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_plots.R
\name{classifierplots_folder}
\alias{classifierplots_folder}
\title{classifier.plots}
\usage{
classifierplots_folder(test.y, pred.prob, folder, height = 5, width = 5)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}

\item{folder}{Directory to save plots into}

\item{height}{height of separately saved plots}

\item{width}{width of separately saved plots}
}
\description{
Produce a suit of classifier diagnostic plots, saving to disk.
}

