% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_plots.R
\name{density.plot}
\alias{density.plot}
\title{density.plot}
\usage{
density.plot(test.y, pred.prob)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}
}
\description{
Returns a ggplot2 plot object containing a score density plot.
If you have trouble displaying, it might be due to the font. Override by passing in something like "Sans"
}

