% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example_predictions}
\alias{example_predictions}
\title{Generated using the gen_example included in the github source}
\description{
Generated using the gen_example included in the github source
}
\keyword{data}

