% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positives.R
\name{positives_plot}
\alias{positives_plot}
\title{positives_plot}
\usage{
positives_plot(test.y, pred.prob)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}
}
\description{
Returns a ggplot2 plot object containing an positives-per-decile plot.
}

