% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_plots.R
\name{precision_plot}
\alias{precision_plot}
\title{precision_plot}
\usage{
precision_plot(test.y, pred.prob, granularity = 0.02, show_numbers = T)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}

\item{granularity}{Default 0.02, probability step between points in plot.}

\item{show_numbers}{Show numbers at deciles T/F default T.}
}
\description{
Returns a ggplot2 plot object containing an precision @ percentile plot
}

