% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_plots.R
\name{precision_thresh_plot}
\alias{precision_thresh_plot}
\title{precision_thresh_plot}
\usage{
precision_thresh_plot(test.y, pred.prob, granularity = 0.02)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}

\item{granularity}{Default 0.02, probability step between points in plot.}
}
\description{
Returns a ggplot2 plot object containing an precision @ threshold plot
}

