% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_plots.R
\name{roc.plot}
\alias{roc.plot}
\title{roc.plot}
\usage{
roc.plot(test.y, pred.prob, use_bootstrap = NULL, smooth = F,
  font = "CM Sans")
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}

\item{use_bootstrap}{Default autodetect. It's really slow for more than say 1k test points, so be aware of this.}

\item{smooth}{Default F. Apply smoothing to the ROC curve. It's a little slow but looks good. Only works with bootstraping on.}

\item{font}{You can override the font here.}
}
\description{
A non-ugly ROC curve! Returns a ggplot2 plot object.
}

