% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_plots.R
\name{sensitivity.thresh.plot}
\alias{sensitivity.thresh.plot}
\title{sensitivity.thresh.plot}
\usage{
sensitivity.thresh.plot(test.y, pred.prob, granularity = 0.02)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}

\item{granularity}{Default 0.02, probability step between points in plot.}
}
\description{
Returns a ggplot2 plot object containing an sensitivity @ threshold plot
}

