% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_plots.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{data}
}
\description{
Logistic sigmoid function, that maps any real number to the [0,1] interval. Supports vectors of numeric.
}

