% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_plots.R
\name{calibration_plot}
\alias{calibration_plot}
\title{calibration_plot}
\usage{
calibration_plot(test.y, pred.prob, granularity = 0.02)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}

\item{granularity}{Default 0.02, probability step between points in plot.}
}
\description{
Returns a ggplot2 plot object containing a smoothed propensity @ prediction level plot
}

