% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc_plot}
\alias{roc_plot}
\title{roc_plot}
\usage{
roc_plot(test.y, pred.prob)
}
\arguments{
\item{test.y}{List of know labels on the test set}

\item{pred.prob}{List of probability predictions on the test set}
}
\description{
A non-ugly ROC curve! Returns a ggplot2 plot object.
}

